common civilian distress file and advertiser adaptative presence
