+++ dat/ai/miner.lua
@@ -1,30 +1,6 @@
 require("dat/ai/tpl/generic.lua")
 require("dat/ai/personality/miner.lua")
-
-
--- Sends a distress signal which causes faction loss
-function sos ()
-   msg = {
-      _("Mayday! We are under attack!"),
-      _("Requesting assistance. We are under attack!"),
-      _("Mining vessel under attack! Requesting help!"),
-      _("Help! Ship under fire!"),
-      _("Taking hostile fire! Need assistance!"),
-      _("We are under attack, require support!"),
-      _("Mayday! Ship taking damage!"),
-      string.format(_("Mayday! Miner %s being assaulted!"), string.lower( ai.pilot():ship():class() ))
-   }
-   ai.settarget( ai.target() )
-   ai.distress( msg[ rnd.int(1,#msg) ])
-end
-
-
-mem.shield_run = 100
-mem.armour_run = 100
-mem.defensive  = false
-mem.enemyclose = 500
-mem.distressmsgfunc = sos
-mem.careful   = true
+require("dat/ai/distress_behaviour.lua")
 
 
 function create ()
